Enable profiling
