+++ comp/core/flare/component.go
@@ -11,6 +11,7 @@ package flare
 
 import (
 	"github.com/DataDog/datadog-agent/comp/core/flare/helpers"
+	"github.com/DataDog/datadog-agent/comp/core/flare/types"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	"go.uber.org/fx"
 )
@@ -20,7 +21,7 @@ import (
 // Component is the component type.
 type Component interface {
 	// Create creates a new flare locally and returns the path to the flare file.
-	Create(pdata ProfileData, ipcError error) (string, error)
+	Create(pdata types.ProfileData, ipcError error) (string, error)
 	// Send sends a flare archive to Datadog.
 	Send(flarePath string, caseID string, email string, source helpers.FlareSource) (string, error)
 }
